Move OpenMode to file_operations module
