+++ app/src/main/java/com/amaze/filemanager/database/CloudHandler.java
@@ -24,8 +24,8 @@ import java.util.List;
 
 import com.amaze.filemanager.database.models.explorer.CloudEntry;
 import com.amaze.filemanager.exceptions.CloudPluginException;
+import com.amaze.filemanager.file_operations.filesystem.OpenMode;
 import com.amaze.filemanager.ui.fragments.CloudSheetFragment;
-import com.amaze.filemanager.utils.OpenMode;
 
 import android.content.Context;
 import android.util.Log;
