+++ app/src/main/java/com/amaze/filemanager/filesystem/RootHelper.java
@@ -25,9 +25,9 @@ import java.util.ArrayList;
 import java.util.List;
 
 import com.amaze.filemanager.exceptions.ShellNotRunningException;
+import com.amaze.filemanager.file_operations.filesystem.OpenMode;
 import com.amaze.filemanager.filesystem.files.FileUtils;
 import com.amaze.filemanager.filesystem.root.ListFilesCommand;
-import com.amaze.filemanager.utils.OpenMode;
 import com.amaze.filemanager.utils.Utils;
 
 import androidx.documentfile.provider.DocumentFile;
