+++ app/src/test/java/com/amaze/filemanager/filesystem/files/FileListSorterTest.java
@@ -34,8 +34,8 @@ import org.robolectric.annotation.Config;
 import org.robolectric.shadows.ShadowDateFormat;
 
 import com.amaze.filemanager.adapters.data.LayoutElementParcelable;
+import com.amaze.filemanager.file_operations.filesystem.OpenMode;
 import com.amaze.filemanager.shadows.ShadowMultiDex;
-import com.amaze.filemanager.utils.OpenMode;
 
 import androidx.test.core.app.ApplicationProvider;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
