+++ app/src/test/java/com/amaze/filemanager/filesystem/ssh/ListFilesOnSshdTest.java
@@ -36,8 +36,8 @@ import java.util.List;
 
 import org.junit.Test;
 
+import com.amaze.filemanager.file_operations.filesystem.OpenMode;
 import com.amaze.filemanager.filesystem.HybridFile;
-import com.amaze.filemanager.utils.OpenMode;
 
 import android.os.Environment;
