+++ app/src/test/java/com/amaze/filemanager/utils/UtilsTest.java
@@ -46,6 +46,7 @@ import org.robolectric.annotation.Config;
 import org.robolectric.shadows.ShadowToast;
 
 import com.amaze.filemanager.R;
+import com.amaze.filemanager.file_operations.filesystem.OpenMode;
 import com.amaze.filemanager.filesystem.HybridFileParcelable;
 
 import android.net.Uri;
